requirements.txt
Flask-RESTful
Flask-SQLAlchemy
FLASK-JWT-Extended